Bit more cleaning
