+++ huff_tests/src/types.rs
@@ -18,6 +18,16 @@ pub enum TestStatus {
     Revert,
 }
 
+/// A test report kind
+pub enum ReportKind {
+    /// Signals `gen_report` to format the test report as a table
+    Table,
+    /// Signals `gen_report` to format the test report as a list
+    List,
+    /// Signals `gen_report` to format the test report as JSON
+    JSON,
+}
+
 /// Convert a TestStatus variant directly to a colored string for use in the report
 impl From<TestStatus> for String {
     fn from(status: TestStatus) -> Self {
@@ -34,6 +44,24 @@ impl From<TestStatus> for Cell {
         match status {
             TestStatus::Success => Cell::new("PASS").fg(Color::Green),
             TestStatus::Revert => Cell::new("FAIL").fg(Color::Red),
+        }
+    }
+}
+
+/// Convert a shared reference to an `Option<String>` to a `ReportKind`.
+/// If the `Option<String>` is `None` or there is no ReportKind match,
+/// `ReportKind::List` is returned.
+impl From<&Option<String>> for ReportKind {
+    fn from(str: &Option<String>) -> Self {
+        if let Some(str) = str {
+            match str.to_lowercase().as_str() {
+                "table" => ReportKind::Table,
+                "list" => ReportKind::List,
+                "json" => ReportKind::JSON,
+                _ => ReportKind::List,
+            }
+        } else {
+            ReportKind::List
         }
     }
 }
